requirements.txt
# Core dependencies for RAG Code Assistant
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.3.7
chromadb==0.4.18
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
streamlit==1.28.1
tiktoken==0.5.1
pathlib==1.0.1
typing-extensions==4.8.0 